+++ firedrake/adjoint_utils/function.py
@@ -1,4 +1,4 @@
-from functools import wraps, cached_property
+from functools import wraps
 import ufl
 from ufl.domain import extract_unique_domain
 from pyadjoint.overloaded_type import create_overloaded_object, FloatingType
@@ -221,7 +221,7 @@ class FunctionMixin(FloatingType):
             return self.copy(deepcopy=True)
 
     def _ad_convert_riesz(self, value, options=None):
-        from firedrake import Function, Cofunction
+        from firedrake import Function
 
         options = {} if options is None else options
         riesz_representation = options.get("riesz_representation", "L2")
@@ -231,14 +231,8 @@ class FunctionMixin(FloatingType):
             # the functional is independent on the control variable.
             return Function(self.function_space())
 
-        if not isinstance(value, (Cofunction, Function)):
-            raise TypeError("Expected a Cofunction or a Function")
-
-        if callable(riesz_representation):
-            return riesz_representation(value)
-
-        return RieszMap(self, riesz_representation,
-                        solver_options=solver_options)(value)
+        return value.riesz_representation(riesz_map=riesz_representation,
+                                          solver_options=solver_options)
 
     @no_annotations
     def _ad_convert_type(self, value, options=None):
@@ -383,118 +377,3 @@ class CofunctionMixin(FunctionMixin):
 
     def _ad_dot(self, other):
         return firedrake.assemble(firedrake.action(self, other))
-
-
-class RieszMap:
-    """Return a map from a dual to a primal function space.
-
-    A `RieszMap` can be called on a `Cofunction` in the appropriate space
-    to yield the `Function` which is the Riesz representer under the.
-
-    Parameters
-    ----------
-    function_space_or_inner_product: FunctionSpace or DualSpace or Function or Cofunction or ufl.Form
-        The space from which to map, or a bilinear form defining an
-        inner product.
-    sobolev_space: String or ufl.sobolevspace.SobolevSpace.
-        Used to determine the inner product.
-    bcs: DirichletBC or list of DirichletBC
-        Boundary conditions to apply to the Riesz map.
-    solver_options: dict
-        A dictionary of PETSc options to be passed to the solver.
-    """
-
-    def __init__(self, function_space_or_inner_product=None,
-                 sobolev_space=ufl.L2, *, bcs=None, solver_options=None):
-        if isinstance(function_space_or_inner_product, ufl.Form):
-            args = ufl.algorithms.extract_arguments(
-                function_space_or_inner_product
-            )
-            if len(args) != 2:
-                raise ValueError(f"inner_product has arity {len(args)}, "
-                                 "should be 2.")
-            function_space = args[0].function_space()
-            inner_product = function_space_or_inner_product
-        else:
-            function_space = function_space_or_inner_product
-            if hasattr(function_space, "function_space"):
-                function_space = function_space.function_space()
-            if ufl.duals.is_dual(function_space):
-                function_space = function_space.dual()
-
-            if str(sobolev_space) == "l2":
-                inner_product = "l2"
-            else:
-                from firedrake import TrialFunction, TestFunction
-                u = TrialFunction(function_space)
-                v = TestFunction(function_space)
-                inner_product = RieszMap._inner_product_form(
-                    sobolev_space, u, v
-                )
-
-        self._function_space = function_space
-        self._inner_product = inner_product
-        self._bcs = bcs
-        self._solver_options = solver_options
-
-    @staticmethod
-    def _inner_product_form(sobolev_space, u, v):
-        from firedrake import inner, dx, grad
-        inner_products = {
-            "L2": lambda u, v: inner(u, v)*dx,
-            "H1": lambda u, v: inner(u, v)*dx + inner(grad(u), grad(v))*dx
-        }
-        try:
-            return inner_products[str(sobolev_space)](u, v)
-        except KeyError:
-            raise ValueError("No inner product defined for Sobolev space "
-                             f"{sobolev_space}.")
-
-    @cached_property
-    def _solver(self):
-        from firedrake import (LinearVariationalSolver,
-                               LinearVariationalProblem, Function, Cofunction)
-        rhs = Cofunction(self._function_space.dual())
-        soln = Function(self._function_space)
-        lvp = LinearVariationalProblem(self._inner_product, rhs, soln,
-                                       bcs=self._bcs)
-        solver = LinearVariationalSolver(
-            lvp, solver_parameters=self._solver_options
-        )
-        return solver.solve, rhs, soln
-
-    def __call__(self, value):
-        """Return the Riesz representer of a Function or Cofunction."""
-        from firedrake import Function, Cofunction
-
-        if ufl.duals.is_dual(value):
-            if value.function_space().dual() != self._function_space:
-                raise ValueError("Function space mismatch in RieszMap.")
-            output = Function(self._function_space)
-
-            if self._inner_product == "l2":
-                for o, c in zip(output.subfunctions, value.subfunctions):
-                    o.dat.data[:] = c.dat.data[:]
-            else:
-                solve, rhs, soln = self._solver
-                rhs.assign(value)
-                solve()
-                output = Function(self._function_space)
-                output.assign(soln)
-        elif ufl.duals.is_primal(value):
-            if value.function_space().dual() != self._function_space:
-                raise ValueError("Function space mismatch in RieszMap.")
-
-            if self._inner_product == "l2":
-                output = Cofunction(self._function_space.dual())
-                for o, c in zip(output.subfunctions, value.subfunctions):
-                    o.dat.data[:] = c.dat.data[:]
-            else:
-                output = firedrake.assemble(
-                    firedrake.action(self._inner_product, value)
-                )
-        else:
-            raise ValueError(
-                f"Unable to ascertain if {value} is primal or dual."
-            )
-        return output
